+++ examples/src/main/kotlin/io/ioUtils.kt
@@ -1,8 +1,11 @@
 package io
 
+import krangl.DataCol
 import krangl.DataFrame
 import krangl.readCSV
+import space.kscience.dataforge.values.Value
 import space.kscience.plotly.Plotly
+import space.kscience.plotly.models.TraceValues
 
 fun readResourceAsString(resource: String): String =
     Plotly.javaClass.getResourceAsStream(resource)?.readAllBytes()?.decodeToString()
@@ -10,3 +13,16 @@ fun readResourceAsString(resource: String): String =
 
 fun readResourceAsCsv(resource: String): DataFrame =
     DataFrame.readCSV(Plotly.javaClass.getResource(resource)?.file?.toString() ?: error("Resource $resource not found"))
+
+/**
+ * Extension function for using krangl data columns as axis values
+ */
+operator fun TraceValues.invoke(column: DataCol) {
+    set(column.values())
+}
+
+fun TraceValues.fromDataFrame(frame: DataFrame, column: String) {
+    invoke(frame[column])
+}
+
+fun DataCol.dfValues(): List<Value> = values().map { Value.of(it) }